requirements.txt
streamlit
pytesseract
Pillow
python-dotenv
langchain-community
langchain-core

